requirements.txt
flask==3.1.1
python-dotenv==1.1.0
langchain==0.3.25
langchain-community==0.3.24
langchain-huggingface==0.1.2
sentence-transformers==4.1.0
faiss-cpu==1.11.0
pypdf==5.6.0
requests==2.32.3
werkzeug==3.1.3